"Add validation for latitude and longitude in Coordinates constructor"
In this commit, the validation for latitude and longitude has been added in the Coordinates constructor. Previously, we had TODO comments to add these validations. Now, when the provided latitude is not between -90 and 90, or longitude is not between -180 and 180, an ArgumentException is immediately thrown. This is necessary to ensure the geocoordinates are always within globally acceptable ranges.